Dot-directory in blog post path no longer excluded
The fourth exclusion pattern tried in skinny/list-published incorrectly excludes any file or directory under a dot-directory. So for example everything under ~/.emacs.d/skinny/blog gets excluded, because of the leading "." in ".emacs.d". (Whether that's a good idea for where to place the skinny docroot is an orthogonal question, I think).

My fix WFM given the inputs I've encountered so far in my (trivial, one-post) setup. I don't really understand what's going on with the trailing "." and ".." filepath suffixes I see coming up in the debugger, which I've guessed are what the fourth pattern is meant to catch, but this fixes the problem for me ATM. I may have misunderstood the intended semantics of the fourth pattern, but there can be no doubting its current implementation is a defect causing the effect I describe. I'm sure you'll want to fix it, using either this fix or something better-informed based on your familiarity with the code.