Prefer "*" to localhost for skinny-host
Using "localhost" may restrict iterative blog composing-and-previewing to a browser running on localhost, or requires port forwarding/tunneling. This is unlikely to be desirable where Emacs is running remotely and editing is done using emacsclient on a remote X display with its own locally-run browser (my own standard scenario). Using "*" removes this restriction.